refactor(period-picker): improved file size and directory structure
