+++ packages/period-picker/src/components/Select.js
