+++ packages/period-picker/src/periodTypes/distinctTypes.js
@@ -1,0 +1,8 @@
+export const DAY = 'day';
+export const WEEK = 'week';
+export const BI_WEEK = 'biWeek';
+export const MONTH = 'month';
+export const BI_MONTH = 'biMonth';
+export const QUARTER = 'quarter';
+export const SIX_MONTH = 'sixMonth';
+export const YEAR = 'year';
